Event callback extensions
Modified the event callback mechanism to support callbacks with a
message argument. This is a breaking change to the event callback
signature. The benefit of providing the full message to the event
callback is that it's no longer necessary to poll all node sensors
when a single sensor value changes or when a node event, like a
battery level update, occurs. The Message class was also modified to
include a reference to the gateway so the handler can use that
information for comparing gateway-specific constants (message types,
etc.).