Don't change sensors dict when loading persistence file
The sensors dict is shared between the gateway and its OTAFirmware
member.

Previous code just loaded a new sensors dict from the persistence file,
thus the gateway and its OTAFirmware members view of the sensors got out
of sync, resulting in that you couldn't ota when loading a persistence
filee.

Signed-off-by: Anton Lundin <glance@acc.umu.se>